pyproject.toml
[tool.poetry]
name = "ray-app"
version = "0.1.0"
description = ""
authors = ["Aybek18 <kasymaliev09@gmail.com>"]
readme = "README.md"
packages = [{include = "ray_app"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
psycopg2-binary = "^2.9.9"
django-environ = "^0.11.2"
pytest = "^7.4.3"
factory-boy = "^3.3.0"
pytest-cov = "^4.1.0"
pytest-django = "^4.7.0"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
huey = "^2.5.0"
django-redis = "^5.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

